Update template (#8)
Update to the snap-in template and some additional documentation. Here
are the details of this update:

1. The reference to docs/airdrop_sdk_library_documentation.md was moved
to Definitions sections so that it's available also to the conformance
tests.

2. In devrev-snapin-template.plain we've added a small implementation
hint that AI consistently got wrong.

3. In docs/airdrop_sdk_library_documentation.md 'function_name' and
'event_type' attributes were missing in AirdropEvent object. This was
really confusing AI so I added it.

4. We are also adding docs/mapping-reasons.mdx that is needed for
initial domain mapping and will be referenced in specific snap-ins (e.g.
Trello). The file was copied from
https://github.com/devrev/fern-api-docs/blob/main/fern/docs/pages/airdrop/data-model/mapping-reasons.mdx

5. From runner.ts I've deleted some unused imports.

Work item: https://app.devrev.ai/devrev/works/ISS-188589