Removing wrong docs update. (#11)
This update to the documentation has been wrong - AirdropEvent doesn't
have function_name and event_type in execution_metadata - thus removing
it.

These two attributes get passed to the runner so it would be very
beneficial for the rendering robustness, if we would document it in the
proper way.

work-item: https://app.devrev.ai/devrev/works/ISS-190744